+++ include/swift/Parse/SyntaxParseActions.h
@@ -82,8 +82,6 @@ struct DeferredNodeInfo {
 // MARK: - SyntaxParseActions
 
 class SyntaxParseActions {
-  llvm::BumpPtrAllocator DeferredNodeAllocator;
-
   virtual void _anchor();
 
 public:
@@ -111,24 +109,24 @@ public:
                                              StringRef leadingTrivia,
                                              StringRef trailingTrivia,
                                              CharSourceRange range,
-                                             bool isMissing);
+                                             bool isMissing) = 0;
 
   /// Create a deferred layout node that may or may not be recorded later using
   /// \c recordDeferredLayout. The \c SyntaxParseAction is responsible for
   /// keeping the deferred token alive until it is destructed.
   virtual OpaqueSyntaxNode
   makeDeferredLayout(syntax::SyntaxKind k, bool isMissing,
-                     const ArrayRef<RecordedOrDeferredNode> &children);
+                     const ArrayRef<RecordedOrDeferredNode> &children) = 0;
 
   /// Record a deferred token node that was previously created using \c
   /// makeDeferredToken. The deferred data will never be used again, so it can
   /// be destroyed by this method. Note that not all deferred nodes will be
   /// recorded and that pending deferred nodes need to be freed when the \c
   /// SyntaxParseActions is destructed.
-  virtual OpaqueSyntaxNode recordDeferredToken(OpaqueSyntaxNode deferred);
+  virtual OpaqueSyntaxNode recordDeferredToken(OpaqueSyntaxNode deferred) = 0;
 
   /// Record a deferred layout node. See recordDeferredToken.
-  virtual OpaqueSyntaxNode recordDeferredLayout(OpaqueSyntaxNode deferred);
+  virtual OpaqueSyntaxNode recordDeferredLayout(OpaqueSyntaxNode deferred) = 0;
 
   /// Since most data of \c ParsedRawSyntax is described as opaque data, the
   /// \c ParsedRawSyntax node needs to reach out to the \c SyntaxParseAction,
@@ -140,11 +138,11 @@ public:
   /// node starts at \p StartLoc.
   virtual DeferredNodeInfo getDeferredChild(OpaqueSyntaxNode node,
                                             size_t childIndex,
-                                            SourceLoc startLoc);
+                                            SourceLoc startLoc) = 0;
 
   /// Return the number of children, \p node has. These can be retrieved using
   /// \c getDeferredChild.
-  virtual size_t getDeferredNumChildren(OpaqueSyntaxNode node);
+  virtual size_t getDeferredNumChildren(OpaqueSyntaxNode node) = 0;
 
   /// Attempt to realize an opaque raw syntax node for a source file into a
   /// SourceFileSyntax node. This will return \c None if the parsing action
