+++ test/Syntax/serialize_tupletype.swift.result
@@ -17,7 +17,7 @@
                 null,
                 null,
                 {
-                  "id": 19,
+                  "id": 1,
                   "tokenKind": {
                     "kind": "kw_typealias"
                   },
@@ -26,7 +26,7 @@
                   "presence": "Present"
                 },
                 {
-                  "id": 20,
+                  "id": 2,
                   "tokenKind": {
                     "kind": "identifier",
                     "text": "x"
@@ -37,11 +37,11 @@
                 },
                 null,
                 {
-                  "id": 18,
+                  "id": 20,
                   "kind": "TypeInitializerClause",
                   "layout": [
                     {
-                      "id": 1,
+                      "id": 3,
                       "tokenKind": {
                         "kind": "equal"
                       },
@@ -50,11 +50,11 @@
                       "presence": "Present"
                     },
                     {
-                      "id": 17,
+                      "id": 19,
                       "kind": "TupleType",
                       "layout": [
                         {
-                          "id": 2,
+                          "id": 4,
                           "tokenKind": {
                             "kind": "l_paren"
                           },
@@ -63,16 +63,16 @@
                           "presence": "Present"
                         },
                         {
-                          "id": 15,
+                          "id": 17,
                           "kind": "TupleTypeElementList",
                           "layout": [
                             {
-                              "id": 9,
+                              "id": 11,
                               "kind": "TupleTypeElement",
                               "layout": [
                                 null,
                                 {
-                                  "id": 3,
+                                  "id": 5,
                                   "tokenKind": {
                                     "kind": "kw__"
                                   },
@@ -81,7 +81,7 @@
                                   "presence": "Present"
                                 },
                                 {
-                                  "id": 4,
+                                  "id": 6,
                                   "tokenKind": {
                                     "kind": "identifier",
                                     "text": "b"
@@ -91,7 +91,7 @@
                                   "presence": "Present"
                                 },
                                 {
-                                  "id": 5,
+                                  "id": 7,
                                   "tokenKind": {
                                     "kind": "colon"
                                   },
@@ -100,11 +100,11 @@
                                   "presence": "Present"
                                 },
                                 {
-                                  "id": 7,
+                                  "id": 9,
                                   "kind": "SimpleTypeIdentifier",
                                   "layout": [
                                     {
-                                      "id": 6,
+                                      "id": 8,
                                       "tokenKind": {
                                         "kind": "identifier",
                                         "text": "Int"
@@ -120,7 +120,7 @@
                                 null,
                                 null,
                                 {
-                                  "id": 8,
+                                  "id": 10,
                                   "tokenKind": {
                                     "kind": "comma"
                                   },
@@ -132,12 +132,12 @@
                               "presence": "Present"
                             },
                             {
-                              "id": 14,
+                              "id": 16,
                               "kind": "TupleTypeElement",
                               "layout": [
                                 null,
                                 {
-                                  "id": 10,
+                                  "id": 12,
                                   "tokenKind": {
                                     "kind": "kw__"
                                   },
@@ -147,7 +147,7 @@
                                 },
                                 null,
                                 {
-                                  "id": 11,
+                                  "id": 13,
                                   "tokenKind": {
                                     "kind": "colon"
                                   },
@@ -156,11 +156,11 @@
                                   "presence": "Present"
                                 },
                                 {
-                                  "id": 13,
+                                  "id": 15,
                                   "kind": "SimpleTypeIdentifier",
                                   "layout": [
                                     {
-                                      "id": 12,
+                                      "id": 14,
                                       "tokenKind": {
                                         "kind": "identifier",
                                         "text": "String"
@@ -183,7 +183,7 @@
                           "presence": "Present"
                         },
                         {
-                          "id": 16,
+                          "id": 18,
                           "tokenKind": {
                             "kind": "r_paren"
                           },
